Remove unnecessary livewire counter and improve code readability
Livewire counter blade was deleted because it was merely used for testing and is no longer needed. The resource and collection files have been updated to return types explicitly for better readability and to follow good coding practices. Code formatting was also improved in multiple files (views, routes, etc.) to adhere to coding conventions and enhance readability.